Fixed unit testing 2 are failing due to a hour conflict likely with timezone
